csv-qna-bot/pyproject.toml
[tool.poetry]
name = "csv-qna-bot"
version = "0.1.0"
description = ""
authors = ["JeongJaeSoon <wjdwotns1006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.274"
pandas = "^2.0.3"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
openai = "^0.27.9"
tabulate = "^0.9.0"
tiktoken = "^0.4.0"
faiss-cpu = "^1.7.4"
slack-bolt = "^1.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
